function_data_ingestion/requirements.txt
google-cloud-workflows==1.5.0
functions-framework
pytest
